requirements.txt
asgiref==3.8.1
Django==4.2.13
django-admin==2.0.2
django-excel-response2==3.0.6
django-filter==24.2
django-six==1.0.5
djangorestframework==3.15.1
drf-yasg==1.21.7
excel-base==1.0.4
importlib_metadata==7.1.0
inflection==0.5.1
isoweek==1.3.3
Markdown==3.6
packaging==24.0
pillow==10.3.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
screen==1.0.1
six==1.16.0
sqlparse==0.5.0
TimeConvert==3.0.13
typing_extensions==4.12.0
tzlocal==5.2
uritemplate==4.1.1
xlwt==1.3.0
zipp==3.18.2
